+++ docs/release-notes.md
@@ -16,6 +16,8 @@ nav_order: 9
 ### Changes
 
 - Warn if template for enabled systemd instance unit has no `Install` section
+- Warn if filesystem overwrites partitioned disk
+- Warn if `wipeTable` overwrites a filesystem that would otherwise be reused
 - Warn if `user`/`group` specified for hard link
 - Install ignition-apply in `/usr/libexec`
 - Convert `NEWS` to Markdown and move to docs site